+++ src/UtilsParsing.cpp
@@ -34,7 +34,7 @@ bool isInt(const string& s) {
 	// allow a negative sign as the first char
 	if (s.at(0) == '-') start=1;
 
-	for (unsigned int i=start; i<s.length(); i++) {
+	for (size_t i=start; i<s.length(); i++) {
 		// if any character in this string is not a numeric digit, this string is not an integer
 		if (s.at(i) < 48 || s.at(i) > 57) return false;
 	}
